fixing demos and sinusoid tones

+++ include/buffer.h
@@ -30,7 +30,7 @@ public:
     std::size_t channels() const;
 
     /* Time in milliseconds associated to a specific frame */
-    unsigned int time(std::size_t frame) const;
+    float time(std::size_t frame) const;
 
     /* Total duration in milliseconds of this sound */
     unsigned int duration() const;
