+++ include/waveform_squared.h
@@ -16,7 +16,7 @@ public:
         WaveformSinusoid(duration, amplitude, frequencyHz)
     {}
     
-    std::int32_t operator()(unsigned int milliseconds) const override
+    std::int32_t operator()(float milliseconds) const override
     {
         std::int32_t sinusoidValue = WaveformSinusoid::operator()(milliseconds);
         return sinusoidValue > 0 ? MAX_INT32 : -MAX_INT32;
